requirements.txt
requests
python-twitter
pyyaml
